+++ src/features/transforms/fields/getField.ts
@@ -1,8 +1,5 @@
 import { ObjectType } from '@features/params/PageParamTypes';
 
-import { CensusData } from '@entities/census/CensusTypes';
-import { KeyboardData } from '@entities/keyboard/KeyboardTypes';
-import { LanguageData } from '@entities/language/LanguageTypes';
 import {
   getCountOfCensuses,
   getCountOfLanguages,
@@ -26,15 +23,20 @@ import {
   getCountOfChildTerritories,
   getCountOfCountries,
 } from '@entities/lib/getObjectRelatedTerritories';
-import { TerritoryData } from '@entities/territory/TerritoryTypes';
 import { ObjectData } from '@entities/types/DataTypes';
-import { WritingSystemData } from '@entities/writingsystem/WritingSystemTypes';
 
 import enforceExhaustiveSwitch from '@shared/lib/enforceExhaustiveness';
 
 import { getLanguageSourcesForEntity } from '../filtering/filterByEnum';
 
 import Field from './Field';
+import {
+  getCensusForEntity,
+  getKeyboardForEntity,
+  getLanguageForEntity,
+  getTerritoryForEntity,
+  getWritingSystemForEntity,
+} from './getEntityConnection';
 
 // Get's a primitive value for a given object and field, used for sorting and filtering.
 // Returns undefined if the field is not applicable to the object type or if the value is missing.
@@ -175,46 +177,6 @@ function getField(object: ObjectData, field: Field): string | number | undefined
     default:
       enforceExhaustiveSwitch(field);
   }
-}
-
-export function getLanguageForEntity(object: ObjectData | undefined): LanguageData | undefined {
-  if (!object) return undefined;
-  if (object.type === ObjectType.Language) return object;
-  if (object.type === ObjectType.Locale) return object.language;
-  if (object.type === ObjectType.Keyboard) return object.language;
-  return undefined;
-}
-
-export function getWritingSystemForEntity(
-  object: ObjectData | undefined,
-): WritingSystemData | undefined {
-  if (!object) return undefined;
-  if (object.type === ObjectType.WritingSystem) return object;
-  if (object.type === ObjectType.Locale) return object.writingSystem;
-  if (object.type === ObjectType.Keyboard) return object.inputWritingSystem;
-  return undefined;
-}
-
-export function getTerritoryForEntity(object: ObjectData | undefined): TerritoryData | undefined {
-  if (!object) return undefined;
-  if (object.type === ObjectType.Territory) return object;
-  if (object.type === ObjectType.Locale) return object.territory;
-  if (object.type === ObjectType.Census) return object.territory;
-  if (object.type === ObjectType.Keyboard) return object.territory;
-  return undefined;
-}
-
-export function getCensusForEntity(object: ObjectData | undefined): CensusData | undefined {
-  if (!object) return undefined;
-  if (object.type === ObjectType.Census) return object;
-  if (object.type === ObjectType.Locale) return object.populationCensus;
-  return undefined;
-}
-
-export function getKeyboardForEntity(object: ObjectData | undefined): KeyboardData | undefined {
-  if (!object) return undefined;
-  if (object.type === ObjectType.Keyboard) return object;
-  return undefined;
 }
 
 export default getField;
